@@ -21,17 +21,17 @@
         }
     ],
     "require": {
-        "php": ">=7.1",
-        "amphp/amp": "^2",
-        "amphp/byte-stream": "^1.6.1",
+        "php": ">=8",
+        "amphp/amp": "^3-dev",
+        "amphp/byte-stream": "^2-dev",
         "amphp/parser": "^1",
-        "amphp/process": "^1",
+        "amphp/process": "^2-dev",
         "amphp/serialization": "^1",
-        "amphp/sync": "^1.0.1"
+        "amphp/sync": "^2-dev"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8 || ^7",
-        "amphp/phpunit-util": "^1.1",
+        "phpunit/phpunit": "^9",
+        "amphp/phpunit-util": "^2-dev",
         "amphp/php-cs-fixer-config": "dev-master"
     },
     "autoload": {
