@@ -5,9 +5,7 @@ namespace Amp\Parallel\Sync;
 use Amp\ByteStream\InputStream;
 use Amp\ByteStream\OutputStream;
 use Amp\ByteStream\StreamException;
-use Amp\Promise;
 use Amp\Serialization\Serializer;
-use function Amp\call;
 
 /**
  * An asynchronous channel for sending data between threads and processes.
@@ -16,17 +14,13 @@ use function Amp\call;
  */
 final class ChannelledStream implements Channel
 {
-    /** @var InputStream */
-    private $read;
+    private InputStream $read;
 
-    /** @var OutputStream */
-    private $write;
+    private OutputStream $write;
 
-    /** @var \SplQueue */
-    private $received;
+    private \SplQueue $received;
 
-    /** @var ChannelParser */
-    private $parser;
+    private ChannelParser $parser;
 
     /**
      * Creates a new channel from the given stream objects. Note that $read and $write can be the same object.
@@ -46,38 +40,34 @@ final class ChannelledStream implements Channel
     /**
      * {@inheritdoc}
      */
-    public function send($data): Promise
+    public function send($data): void
     {
-        return call(function () use ($data): \Generator {
-            try {
-                return yield $this->write->write($this->parser->encode($data));
-            } catch (StreamException $exception) {
-                throw new ChannelException("Sending on the channel failed. Did the context die?", 0, $exception);
-            }
-        });
+        try {
+            $this->write->write($this->parser->encode($data));
+        } catch (StreamException $exception) {
+            throw new ChannelException("Sending on the channel failed. Did the context die?", 0, $exception);
+        }
     }
 
     /**
      * {@inheritdoc}
      */
-    public function receive(): Promise
+    public function receive(): mixed
     {
-        return call(function (): \Generator {
-            while ($this->received->isEmpty()) {
-                try {
-                    $chunk = yield $this->read->read();
-                } catch (StreamException $exception) {
-                    throw new ChannelException("Reading from the channel failed. Did the context die?", 0, $exception);
-                }
-
-                if ($chunk === null) {
-                    throw new ChannelException("The channel closed unexpectedly. Did the context die?");
-                }
-
-                $this->parser->push($chunk);
+        while ($this->received->isEmpty()) {
+            try {
+                $chunk = $this->read->read();
+            } catch (StreamException $exception) {
+                throw new ChannelException("Reading from the channel failed. Did the context die?", 0, $exception);
             }
 
-            return $this->received->shift();
-        });
+            if ($chunk === null) {
+                throw new ChannelException("The channel closed unexpectedly. Did the context die?");
+            }
+
+            $this->parser->push($chunk);
+        }
+
+        return $this->received->shift();
     }
 }
