@@ -4,11 +4,12 @@ namespace Amp\Parallel\Test\Sync;
 
 use Amp\Delayed;
 use Amp\Parallel\Context\Process;
+use Amp\Parallel\Sync\Parcel;
 use Amp\Parallel\Sync\SharedMemoryException;
 use Amp\Parallel\Sync\SharedMemoryParcel;
-use Amp\Promise;
-use Amp\Success;
 use Amp\Sync\SyncException;
+use function Amp\async;
+use function Amp\await;
 
 /**
  * @requires extension shmop
@@ -18,12 +19,12 @@ class SharedMemoryParcelTest extends AbstractParcelTest
 {
     const ID = __CLASS__;
 
-    private $parcel;
+    private ?SharedMemoryParcel $parcel;
 
-    protected function createParcel($value): Promise
+    protected function createParcel($value): Parcel
     {
         $this->parcel = SharedMemoryParcel::create(self::ID, $value);
-        return new Success($this->parcel);
+        return $this->parcel;
     }
 
     public function tearDown(): void
@@ -34,7 +35,7 @@ class SharedMemoryParcelTest extends AbstractParcelTest
     public function testObjectOverflowMoved(): \Generator
     {
         $object = SharedMemoryParcel::create(self::ID, 'hi', 2);
-        yield $object->synchronized(function () {
+        $object->synchronized(function () {
             return 'hello world';
         });
 
@@ -51,18 +52,18 @@ class SharedMemoryParcelTest extends AbstractParcelTest
 
         $process = new Process([__DIR__ . '/Fixture/parcel.php', self::ID]);
 
-        $promise = $object->synchronized(function (int $value): \Generator {
+        $promise = async(fn() => $object->synchronized(function (int $value): \Generator {
             $this->assertSame(42, $value);
             yield new Delayed(500); // Child must wait until parent finishes with parcel.
             return $value + 1;
-        });
+        }));
 
-        yield $process->start();
+        $process->start();
 
-        $this->assertSame(43, yield $promise);
+        $this->assertSame(43, await($promise));
 
-        $this->assertSame(44, yield $process->join()); // Wait for child process to finish.
-        $this->assertEquals(44, yield $object->unwrap());
+        $this->assertSame(44, $process->join()); // Wait for child process to finish.
+        $this->assertEquals(44, $object->unwrap());
     }
 
     public function testInvalidSize(): void
