@@ -5,13 +5,13 @@
 
 use Amp\Parallel\Sync\Channel;
 
-return function (Channel $channel): \Generator {
-    \printf("Received the following from parent: %s\n", yield $channel->receive());
+return function (Channel $channel): int {
+    \printf("Received the following from parent: %s\n", $channel->receive());
 
     print "Sleeping for 3 seconds...\n";
     \sleep(3); // Blocking call in process.
 
-    yield $channel->send("Data sent from child.");
+    $channel->send("Data sent from child.");
 
     print "Sleeping for 2 seconds...\n";
     \sleep(2); // Blocking call in process.
