@@ -6,6 +6,8 @@ require \dirname(__DIR__) . '/vendor/autoload.php';
 use Amp\Loop;
 use Amp\Parallel\Worker\CallableTask;
 use Amp\Parallel\Worker\DefaultPool;
+use function Amp\async;
+use function Amp\await;
 
 // A variable to store our fetched results
 $results = [];
@@ -18,28 +20,23 @@ $tasks = [
 ];
 
 // Event loop for parallel tasks
-Loop::run(function () use (&$results, $tasks) {
-    $timer = Loop::repeat(200, function () {
-        \printf(".");
-    });
-    Loop::unreference($timer);
-
-    $pool = new DefaultPool;
-
-    $coroutines = [];
-
-    foreach ($tasks as $index => $task) {
-        $coroutines[] = Amp\call(function () use ($pool, $index, $task) {
-            $result = yield $pool->enqueue($task);
-            \printf("\nRead from task %d: %d bytes\n", $index, \strlen($result));
-            return $result;
-        });
-    }
-
-    $results = yield Amp\Promise\all($coroutines);
-
-    return yield $pool->shutdown();
+$timer = Loop::repeat(200, function () {
+    \printf(".");
 });
+Loop::unreference($timer);
 
-echo "\nResult array keys:\n";
-echo \var_export(\array_keys($results), true);
+$pool = new DefaultPool;
+
+$promises = [];
+
+foreach ($tasks as $index => $task) {
+    $promises[] = async(function () use ($pool, $index, $task): string {
+        $result = $pool->enqueue($task);
+        \printf("\nRead from task %d: %d bytes\n", $index, \strlen($result));
+        return $result;
+    });
+}
+
+$results = await($promises);
+
+$pool->shutdown();
