@@ -3,7 +3,6 @@
 namespace Amp\Parallel\Context;
 
 use Amp\Loop;
-use Amp\Promise;
 
 const LOOP_FACTORY_IDENTIFIER = ContextFactory::class;
 
@@ -13,23 +12,9 @@ const LOOP_FACTORY_IDENTIFIER = ContextFactory::class;
  *
  * @return Context
  */
-function create($script): Context
+function create(string|array $script): Context
 {
     return factory()->create($script);
-}
-
-/**
- * Creates and starts a process based on installed extensions (a thread if ext-parallel is installed, otherwise a child
- * process).
- *
- * @param string|string[] $script Path to PHP script or array with first element as path and following elements options
- *     to the PHP script (e.g.: ['bin/worker', 'Option1Value', 'Option2Value'].
- *
- * @return Promise<Context>
- */
-function run($script): Promise
-{
-    return factory()->run($script);
 }
 
 /**
