@@ -16,22 +16,22 @@ interface Environment extends \ArrayAccess
      *
      * @return mixed|null Returns null if the key does not exist.
      */
-    public function get(string $key);
+    public function get(string $key): mixed;
 
     /**
      * @param string $key
      * @param mixed $value Using null for the value deletes the key.
      * @param int $ttl Number of seconds until data is automatically deleted. Use null for unlimited TTL.
      */
-    public function set(string $key, $value, int $ttl = null);
+    public function set(string $key, $value, int $ttl = null): void;
 
     /**
      * @param string $key
      */
-    public function delete(string $key);
+    public function delete(string $key): void;
 
     /**
      * Removes all values.
      */
-    public function clear();
+    public function clear(): void;
 }
