@@ -9,7 +9,7 @@ use Amp\Parallel\Context\Parallel;
  */
 final class WorkerParallel extends TaskWorker
 {
-    const SCRIPT_PATH = __DIR__ . "/Internal/worker-process.php";
+    private const SCRIPT_PATH = __DIR__ . "/Internal/worker-process.php";
 
     /**
      * @param string $envClassName Name of class implementing \Amp\Parallel\Worker\Environment to instigate.
@@ -18,7 +18,7 @@ final class WorkerParallel extends TaskWorker
      *
      * @throws \Error If the PHP binary path given cannot be found or is not executable.
      */
-    public function __construct(string $envClassName = BasicEnvironment::class, string $bootstrapPath = null)
+    public function __construct(string $envClassName = BasicEnvironment::class, ?string $bootstrapPath = null)
     {
         $script = [
             self::SCRIPT_PATH,
