@@ -2,9 +2,9 @@
 
 namespace Amp\Parallel\Test\Worker;
 
-use Amp\Delayed;
 use Amp\Parallel\Worker\BasicEnvironment;
 use Amp\PHPUnit\AsyncTestCase;
+use function Amp\delay;
 
 class BasicEnvironmentTest extends AsyncTestCase
 {
@@ -33,7 +33,8 @@ class BasicEnvironmentTest extends AsyncTestCase
         $environment = new BasicEnvironment;
         $key = "key";
 
-        $this->assertNull($environment->set($key, null));
+        $environment->set($key, null);
+        $this->assertFalse($environment->exists($key));
     }
 
     public function testSetShouleThrowError()
@@ -86,7 +87,7 @@ class BasicEnvironmentTest extends AsyncTestCase
 
         $environment->set($key, 1, 2);
 
-        yield new Delayed(3000);
+        delay(3000);
 
         $this->assertFalse($environment->exists($key));
     }
@@ -103,7 +104,7 @@ class BasicEnvironmentTest extends AsyncTestCase
 
         $environment->set($key, 2);
 
-        yield new Delayed(2000);
+        delay(2000);
 
         $this->assertTrue($environment->exists($key));
         $this->assertSame(2, $environment->get($key));
@@ -117,7 +118,7 @@ class BasicEnvironmentTest extends AsyncTestCase
         $environment->set($key, 1, 10);
         $environment->set($key, 1, 1);
 
-        yield new Delayed(2000);
+        delay(2000);
 
         $this->assertFalse($environment->exists($key));
     }
@@ -130,11 +131,11 @@ class BasicEnvironmentTest extends AsyncTestCase
         $environment->set($key, 1, 1);
         $environment->set($key, 1, 3);
 
-        yield new Delayed(2000);
+        delay(2000);
 
         $this->assertTrue($environment->exists($key));
 
-        yield new Delayed(1100);
+        delay(1100);
 
         $this->assertFalse($environment->exists($key));
     }
@@ -148,14 +149,17 @@ class BasicEnvironmentTest extends AsyncTestCase
         $environment->set($key1, 1, 2);
         $environment->set($key2, 2, 2);
 
-        yield new Delayed(1000);
+        delay(1000);
 
         $this->assertSame(1, $environment->get($key1));
         $this->assertTrue($environment->exists($key2));
 
-        yield new Delayed(1500);
+        delay(1500);
 
         $this->assertTrue($environment->exists($key1));
         $this->assertFalse($environment->exists($key2));
+
+        $environment->delete($key1);
+        delay(1000); // Let TTL watcher deactivate itself
     }
 }
