@@ -3,6 +3,7 @@
 namespace Amp\Parallel\Test\Worker;
 
 use Amp\Parallel\Worker\BasicEnvironment;
+use Amp\Parallel\Worker\Worker;
 use Amp\Parallel\Worker\WorkerParallel;
 
 /**
@@ -10,7 +11,7 @@ use Amp\Parallel\Worker\WorkerParallel;
  */
 class WorkerParallelTest extends AbstractWorkerTest
 {
-    protected function createWorker(string $envClassName = BasicEnvironment::class, string $autoloadPath = null)
+    protected function createWorker(string $envClassName = BasicEnvironment::class, string $autoloadPath = null): Worker
     {
         return new WorkerParallel($envClassName, $autoloadPath);
     }
